+++ toggleselection/__init__.py
@@ -4,8 +4,8 @@
 from fman import DirectoryPaneCommand, DirectoryPaneListener, load_json, save_json, PLATFORM
 from core.commands.util import is_hidden
 from fman.url import splitscheme
-import json
-from statusbarextended import StatusBarExtended
+import statusbarextended        as SBE
+import statusbarextended_config as SBEcfg
 
 class _CorePaneCommand(DirectoryPaneCommand): # copy from core/commands/__init__.py
     def select_all(self):
@@ -81,12 +81,15 @@ class SelectionOverride(DirectoryPaneListener):
             'move_cursor_page_down', 'move_cursor_page_up',
             'move_cursor_home'     , 'move_cursor_end'):
             getattr(_CorePaneCommand, command_name)(self, args)
-            self.show_selected_files()
+            if 'toggle_selection' in args:
+                if args['toggle_selection'] == True:
+                    self.show_selected_files()
             return 'command_empty', {}
 
     def show_selected_files(self):
-        statusBarExtendedEnabled = load_json('StatusBarExtended.json')
-        if statusBarExtendedEnabled:
-            statusBarExtendedEnabledJson = json.loads(statusBarExtendedEnabled)
-            if statusBarExtendedEnabledJson['enabled'] == True:
-                StatusBarExtended.show_selected_files(self)
+        cfg = SBEcfg.SingletonConfig()
+        cfgCurrent, exit_status = cfg.loadConfig()
+        if  cfgCurrent is None:
+            return
+        if  cfgCurrent["Enabled"] == True:
+            SBE.StatusBarExtended.show_selected_files(self, cfgCurrent)